separate-interfaces
This change is to break interfaces out of the classes, preparing the package
to be used within an OSGi bundle.
